Rename libyui_clint to yui_rest_client
