+++ Gemfile
@@ -2,7 +2,7 @@
 
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in libyui_client.gemspec
+# Specify your gem's dependencies in yui_rest_client.gemspec
 gemspec
 gem 'rubocop'
 gem 'webmock'
