--- lib/libyui_client/widgets.rb
@@ -1,30 +1,0 @@
-# frozen_string_literal: true
-
-module LibyuiClient
-  module Widgets
-    require 'libyui_client/waitable'
-    require 'libyui_client/widgets/base'
-    require 'libyui_client/widgets/bargraph'
-    require 'libyui_client/widgets/button'
-    require 'libyui_client/widgets/checkbox'
-    require 'libyui_client/widgets/combobox'
-    require 'libyui_client/widgets/datefield'
-    require 'libyui_client/widgets/label'
-    require 'libyui_client/widgets/menubutton'
-    require 'libyui_client/widgets/multilinebox'
-    require 'libyui_client/widgets/numberbox'
-    require 'libyui_client/widgets/progressbar'
-    require 'libyui_client/widgets/radiobutton'
-    require 'libyui_client/widgets/selectionbox'
-    require 'libyui_client/widgets/tab'
-    require 'libyui_client/widgets/table'
-    require 'libyui_client/widgets/textbox'
-    require 'libyui_client/widgets/timefield'
-    require 'libyui_client/widgets/label'
-    require 'libyui_client/widgets/richtext'
-    require 'libyui_client/widgets/tree'
-    # Class representing a Wizard UI. It can be YWizard
-    class Wizard < Widgets::Base
-    end
-  end
-end
