+++ lib/yui_rest_client/actions.rb
