+++ lib/yui_rest_client/app.rb
