+++ lib/yui_rest_client/error.rb
