+++ lib/yui_rest_client/filter_extractor.rb
