+++ lib/yui_rest_client/http/http_client.rb
