+++ lib/yui_rest_client/http/response.rb
