+++ lib/yui_rest_client/http/version_controller.rb
