+++ lib/yui_rest_client/http/widget_controller.rb
