+++ lib/yui_rest_client/local_process.rb
