+++ lib/yui_rest_client/logger.rb
