+++ lib/yui_rest_client/timer.rb
