+++ lib/yui_rest_client/version.rb
