+++ lib/yui_rest_client/wait.rb
