+++ lib/yui_rest_client/waitable.rb
