+++ lib/yui_rest_client/widgets.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+module LibyuiClient
+  module Widgets
+    require 'yui_rest_client/waitable'
+    require 'yui_rest_client/widgets/base'
+    require 'yui_rest_client/widgets/bargraph'
+    require 'yui_rest_client/widgets/button'
+    require 'yui_rest_client/widgets/checkbox'
+    require 'yui_rest_client/widgets/combobox'
+    require 'yui_rest_client/widgets/datefield'
+    require 'yui_rest_client/widgets/label'
+    require 'yui_rest_client/widgets/menubutton'
+    require 'yui_rest_client/widgets/multilinebox'
+    require 'yui_rest_client/widgets/numberbox'
+    require 'yui_rest_client/widgets/progressbar'
+    require 'yui_rest_client/widgets/radiobutton'
+    require 'yui_rest_client/widgets/selectionbox'
+    require 'yui_rest_client/widgets/tab'
+    require 'yui_rest_client/widgets/table'
+    require 'yui_rest_client/widgets/textbox'
+    require 'yui_rest_client/widgets/timefield'
+    require 'yui_rest_client/widgets/label'
+    require 'yui_rest_client/widgets/richtext'
+    require 'yui_rest_client/widgets/tree'
+    # Class representing a Wizard UI. It can be YWizard
+    class Wizard < Widgets::Base
+    end
+  end
+end
