+++ lib/yui_rest_client/widgets/bargraph.rb
