+++ lib/yui_rest_client/widgets/base.rb
