+++ lib/yui_rest_client/widgets/checkbox.rb
