+++ lib/yui_rest_client/widgets/combobox.rb
