+++ lib/yui_rest_client/widgets/label.rb
