+++ lib/yui_rest_client/widgets/menubutton.rb
