+++ lib/yui_rest_client/widgets/multilinebox.rb
