+++ lib/yui_rest_client/widgets/numberbox.rb
