+++ lib/yui_rest_client/widgets/progressbar.rb
