+++ lib/yui_rest_client/widgets/radiobutton.rb
