+++ lib/yui_rest_client/widgets/richtext.rb
