+++ lib/yui_rest_client/widgets/selectionbox.rb
