+++ lib/yui_rest_client/widgets/tab.rb
