+++ lib/yui_rest_client/widgets/table.rb
