+++ lib/yui_rest_client/widgets/textbox.rb
