+++ lib/yui_rest_client/widgets/tree.rb
